Don't move Firefox::General bugs to Core::Internationalization
Bugbug wrongly learned that the bugs mentioning Fluent should go in Core::Internationalization
because over the past few months all bugs mentioning Fluent are in that component (while
Fluent was being built and starting to be used).
Now that Fluent is ready, bugs to track migrations are no longer filed in that component, but are
filed in the component of the migration.
See https://github.com/mozilla/bugbug/issues/1484.