Automatically move spambugs into the Invalid Bugs product and close
Automatically move bugs where we have very high confidence in them being
spam/invalid into the "Invalid Bugs" product and mark them as
RESOLVED/INVALID.

The comment added to the bug includes the confidence score, instructions
to revert the change if this is in error, as well as a generic warning
not to use BMO for test bugs.

The required confidence score has been increased from 90% to 95% in
order to reduce false positives.