Set webcompat:platform-bug on site reports likely to be platform bugs
If an webcompat site report bug is:
* Blocked on an open platform bug
* Not in the needs:diagnosis state
* Does not have the webcompat:platform-bug keyword, and has never had
that keyword in the past.

Then it is probably supposed to have the webcompat:platform-bug keyword
as it fits the criteria for being considered a webcompat bug likley to
be fixed by a platform change.

We don't include bugs where the keyword was previously removed
because there are instances where a platform fix would offer a partial
solution, but is not on the critical path to a full fix (e.g. if the
site is blocking firefox in part due to some feature that's easy to
avoid or polyfill).

Also, to avoid a potential race condition, we only update bugs if they
haven't themselves been updated since the last import of the bugzilla
dat ainto BigQuery.