Add a rule to add See Also links to bugs tagged with web-features
We generally want one platform bug for each web-feature we might
implement. Once we know the web-feature associated with the bug, we
can automatically ensure that we link to a lot of data about the
feature, including the entry in the web-features-explorer and the spec
URL. In some cases we also know the standards position, and can also
ensure that's linked in the See Also field.

For simplicity the linked data we pull from BigQuery uses full string
matches on the URLs, but for actual linking we just compare host and
pathnames, so we don't duplicate links with different fragments or similar.