Update webcompat_sightline rule to only query bugs that will change
This was failing because bugzilla was rejecting the query.

Since we now have the whiteboard data in BigQuery we can change the
implementation to get a list of bug ids that had incorrect data at the
time of last BQ import, and only query those bugs from bugzilla. This
makes the bugzilla query much shorter, and so less likely to run into
API limits.