Add [webcompat:japan] to whiteboard for bugs in Japan Top 1000
This updates the WebcompatSightline rule to be more generic and allow
providing labels corresponding to multiple subsets of the webcompat
bugs that we're interested in.

Concretely it adds the capability to label bugs in the Japan Top 1000
as `[webcompat:japan]` in the whiteboard.

To avoid creating over-long bugzilla queries, rather than selecting
all bugs that are in the various metrics by number, we select from the
database only the bugs where the last recorded whiteboard value
doesn't match the expectations given its metric status. This is
conservative: although the bugzilla view of the whiteboard may be
outdated, we check again against the current value before making any
changes.