Add user-impact-score to core bugs that block webcompat site reports
Query for user impact score on webcompat core bugs, and add this to
the user story, as we do for site reports.

For now we don't set the cf_webcompat_score field for these bugs, as
we need to decide if that's useful and whether to use the same
score boundaries as for the individual site reports.